ca: silence warning

+++ modules/ca/src/client/tcpiiu.cpp
@@ -1058,7 +1058,7 @@ void tcpiiu::show ( unsigned level ) const
             this->_receiveThreadIsBusy );
     }
     if ( level > 2u ) {
-        ::printf ( "\tvirtual circuit socket identifier %d\n", this->sock );
+        ::printf ( "\tvirtual circuit socket identifier %d\n", (int)this->sock );
         ::printf ( "\tsend thread flush signal:\n" );
         this->sendThreadFlushEvent.show ( level-2u );
         ::printf ( "\tsend thread:\n" );
